requirements.txt
requests>=2.28
yt-dlp>=2024.5.20
python-dateutil>=2.8
